+++ biblioteca-4/comunes/biblioteca.php
@@ -372,7 +372,7 @@ function pie()
     print "  <footer>\n";
     print "    <p class=\"ultmod\">\n";
     print "      Última modificación de esta página:\n";
-    print "      <time datetime=\"2021-05-06\">6 de mayo de 2021</time>\n";
+    print "      <time datetime=\"2021-05-10\">10 de mayo de 2021</time>\n";
     print "    </p>\n";
     print "\n";
     print "    <p class=\"licencia\">\n";