+++ biblioteca-4/tablas/db-prestamos/modificar-3.php
@@ -27,6 +27,8 @@ if (hayErrores("modificar-3")) {
 
 [$id_persona, $id_obra, $prestado, $devuelto, $id] = compruebaAvisosIndividuales("modificar-3", "prestamos", "id_persona", "id_obra", "prestado", "devuelto", "id");
 
+incluyeValoresOriginalesEnAvisos("modificar-3", "prestamos", "id_persona", "id_obra", "prestado", "devuelto", "id");
+
 compruebaAvisosGenerales("modificar-3", "todosVaciosMenosPrimero", "id", "id_persona", "id_obra", "prestado", "devuelto");
 
 compruebaAvisosGenerales("modificar-3", "yaExisteRegistroConOtroId", "prestamos", "id_persona", "id_obra", "prestado", "devuelto", "id");